requirements.txt
Flask==0.10.1
PyYAML==3.11
WebTest==2.0.18
factory-boy==2.5.0
nose==1.3.4
pinocchio==0.4.2
preggy==1.2.0
#requests==2.5.3
redis==2.10.3
uWSGI==2.0.10
python-gcm==0.2
validator.py==1.2
couchdb==1.0
gearman==2.0.2
peewee==2.6.3
psycopg2==2.6
pyexcel==0.1.7
AlphaID
carrot==0.10.7
pika==0.9.8
#pycouchdb==1.13
elasticsearch
flask-paginate